+++ tests/abc_test.py
@@ -1,19 +1,21 @@
 import pytest
 from torch import eye, ones, zeros
+from torch.distributions import MultivariateNormal
 
+from sbi.inference import ABC, SMC
 from sbi.simulators.linear_gaussian import (
     linear_gaussian,
     true_posterior_linear_gaussian_mvn_prior,
 )
-
-from sbi.inference import ABC, SMC
-from torch.distributions import MultivariateNormal
 from tests.test_utils import check_c2st
 
 
 @pytest.mark.parametrize("num_dim", (1, 2))
 def test_mcabc_inference_on_linear_gaussian(
-    num_dim, lra=False, sass=False, sass_expansion_degree=1,
+    num_dim,
+    lra=False,
+    sass=False,
+    sass_expansion_degree=1,
 ):
     x_o = zeros((1, num_dim))
     num_samples = 1000
@@ -54,7 +56,10 @@ def test_mcabc_inference_on_linear_gaussian(
 def test_mcabc_sass_lra(lra, sass_expansion_degree, set_seed):
 
     test_mcabc_inference_on_linear_gaussian(
-        num_dim=2, lra=lra, sass=True, sass_expansion_degree=sass_expansion_degree,
+        num_dim=2,
+        lra=lra,
+        sass=True,
+        sass_expansion_degree=sass_expansion_degree,
     )
 
 
@@ -85,7 +90,7 @@ def test_smcabc_inference_on_linear_gaussian(
         num_particles=1000,
         num_initial_pop=5000,
         epsilon_decay=0.5,
-        num_simulations=100000,
+        num_simulations=110000,
         distance_based_decay=True,
         lra=lra,
         sass=sass,