+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## Minor changes
 - Depend on new `joblib=1.0.0` and fix progress bar updates for multiprocessing (#421).
 - Fix for embedding nets with `SNRE` (thanks @adittmann, #425)
+- Is it now optional to pass a prior distribution when using SNPE (#426)
 
 
 # v0.14.3